+++ lib/ui/settings/SettingsMealPlanPanel.dart
@@ -224,7 +224,7 @@ class _SettingsMealPlanPanelState extends State<SettingsMealPlanPanel> implement
 
   Widget _buildMealPlanSection() {
     bool isSignedIn = Auth().isLoggedIn;
-    List<Widget> widgets = List();
+    List<Widget> widgets = [];
     widgets.add(Padding(padding: EdgeInsets.only(top: 16)));
     if (!isSignedIn) {
       widgets.add(Padding(
@@ -262,7 +262,7 @@ class _SettingsMealPlanPanelState extends State<SettingsMealPlanPanel> implement
             Expanded(
               child: VerticalTitleContentSection(
                 title: Localization().getStringEx(
-                    "panel.settings.meal_plan.label.cafe_credits.text", "Café Credits"),
+                    "panel.settings.meal_plan.label.dining_dollars.text", "Dining Dollars"),
                 content: IlliniCash().ballance?.cafeCreditBalanceDisplayText ?? "0",
               ),
             )
@@ -383,10 +383,10 @@ class _SettingsMealPlanPanelState extends State<SettingsMealPlanPanel> implement
           'panel.settings.meal_plan.label.description', 'Description');
       String amountHeader = Localization().getStringEx(
           'panel.settings.meal_plan.label.amount', 'Amount');
-      List<Widget> dateWidgets = new List();
-      List<Widget> locationWidgets = new List();
-      List<Widget> descriptionWidgets = new List();
-      List<Widget> amountViewWidgets = new List();
+      List<Widget> dateWidgets =  [];
+      List<Widget> locationWidgets =  [];
+      List<Widget> descriptionWidgets =  [];
+      List<Widget> amountViewWidgets =  [];
       //Headers
       dateWidgets.add(_buildBalanceTableHeaderItem(dateHeader));
       locationWidgets.add(_buildBalanceTableHeaderItem(locationHeader));
@@ -551,7 +551,7 @@ class _SettingsMealPlanPanelState extends State<SettingsMealPlanPanel> implement
             Row(
               mainAxisAlignment: MainAxisAlignment.end,
               children: <Widget>[
-                FlatButton(
+                TextButton(
                     onPressed: () {
                       Analytics.instance.logAlert(text:text, selection: "Ok");
                       Navigator.pop(context);