+++ lib/ui/settings/SettingsNotificationsPanel.dart
@@ -106,7 +106,7 @@ class _SettingsNotificationsPanelState extends State<SettingsNotificationsPanel>
   Widget _buildSettings(){
     BorderRadius _bottomRounding = BorderRadius.only(bottomLeft: Radius.circular(5), bottomRight: Radius.circular(5));
     BorderRadius _topRounding = BorderRadius.only(topLeft: Radius.circular(5), topRight: Radius.circular(5));
-    List<Widget> widgets = List();
+    List<Widget> widgets = [];
 
     widgets.add(_CustomToggleButton(
           enabled: _notificationsEnabled,