+++ lib/ui/widgets/MapWidget.dart
@@ -77,7 +77,7 @@ class MapController {
     var options = {
       "LocationThresoldDistance": Storage().debugMapThresholdDistance
     };
-    List<dynamic> jsonData = List<dynamic>();
+    List<dynamic> jsonData = [];
     if (AppCollection.isCollectionNotEmpty(explores)) {
       for (dynamic explore in explores) {
         jsonData.add(explore.toJson());