+++ CHANGELOG.md
@@ -6,6 +6,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Fixed
+- Handled location error in iOS Directions controller [#254](https://github.com/rokwire/safer-illinois-app/issues/254).
+
+## [2.4.11] - 2021-05-26
+### Fixed
+- Illini Cash isn't displaying unless logged in [#252](https://github.com/rokwire/safer-illinois-app/issues/252).
+
+## [2.4.10] - 2021-05-21
+### Changed
+- Name change - "Cafe Credits" to "Dining Dollars" [#250](https://github.com/rokwire/safer-illinois-app/issues/250).
+
+## [2.4.9] - 2021-05-18
+### Changed
+- Flutter 2.0 integration [#245](https://github.com/rokwire/safer-illinois-app/issues/245).
+
+## [2.4.8] - 2021-05-14
+### Added
+- Added capability to override Styles content from Settings Debug panel [#246](https://github.com/rokwire/safer-illinois-app/issues/246).
+
+## [2.4.7] - 2021-05-13
+### Fixed
 - Handled exceptions when system date time is much behind the current date time [#243](https://github.com/rokwire/safer-illinois-app/issues/243).
 
 ## [2.4.6] - 2021-04-12